Extract super class CodeModelGenerator.
